VFS: Migrate to 64-bit inode numbers and size_t for read/write
- Changed fs_node_t.inode from uint32_t to uint64_t
- Changed struct dirent.ino from uint32_t to uint64_t
- Changed read/write callbacks to use size_t instead of uint32_t
- Changed readdir index from uint32_t to uint64_t
- Added const qualifier to write buffer parameter
- Updated vfs.c wrapper functions to match new signatures
- Updated ext2 driver callbacks to match new VFS signatures

This is an ABI-breaking change but provides proper support for
large filesystems and file sizes.

+++ sys/fs/ext2/ext2.c
@@ -17,8 +17,8 @@ static int ext2_node_cache_idx = 0;
 
 // Forward declarations
 static fs_node_t *ext2_mount(const char *device, uint32_t flags, void *data);
-static uint32_t ext2_file_read(fs_node_t *node, off_t offset, uint32_t size, uint8_t *buffer);
-static struct dirent *ext2_readdir(fs_node_t *node, uint32_t index);
+static size_t ext2_file_read(fs_node_t *node, off_t offset, size_t size, uint8_t *buffer);
+static struct dirent *ext2_readdir(fs_node_t *node, uint64_t index);
 static fs_node_t *ext2_finddir(fs_node_t *node, char *name);
 static int ext2_readlink_fn(fs_node_t *node, char *buf, size_t size);
 
@@ -328,13 +328,13 @@ static int ext2_readlink_fn(fs_node_t *node, char *buf, size_t size) {
 }
 
 // File read operation
-static uint32_t ext2_file_read(fs_node_t *node, off_t offset, uint32_t size, uint8_t *buffer) {
+static size_t ext2_file_read(fs_node_t *node, off_t offset, size_t size, uint8_t *buffer) {
     ext2_node_t *ctx = (ext2_node_t *)(uintptr_t)node->impl;
     return ext2_inode_read(ctx->fs, &ctx->inode, offset, size, buffer);
 }
 
 // File write operation
-static uint32_t ext2_file_write(fs_node_t *node, off_t offset, uint32_t size, const uint8_t *buffer) {
+static size_t ext2_file_write(fs_node_t *node, off_t offset, size_t size, const uint8_t *buffer) {
     ext2_node_t *ctx = (ext2_node_t *)(uintptr_t)node->impl;
     ext2_fs_t *fs = ctx->fs;
     
@@ -354,7 +354,7 @@ static struct dirent ext2_dirent;
 static uint8_t ext2_dir_buf[4096];
 
 // Read directory entry at index
-static struct dirent *ext2_readdir(fs_node_t *node, uint32_t index) {
+static struct dirent *ext2_readdir(fs_node_t *node, uint64_t index) {
     ext2_node_t *ctx = (ext2_node_t *)(uintptr_t)node->impl;
     ext2_fs_t *fs = ctx->fs;
